Updated and refactored data access and coefficient decoding sections. Spec should be up to date and accurate for coefficient decoding, but fine-tuning and checking is required. Diagrams also needed.
Major change is in the way arithmetic decoding operations are called. Just
specify sets of contexts to be used for follow bits, data bits and the sign
by passing indices into the state variable, which maintains the contexts.

This simplifies the code in the stream syntax, avoiding local copies etc.
Explanation will be required for the context indices. The arithmetic decoding
section needs updating to reflect this change and use of interleaved exp-
Golomb coding.

SKIPPED:
	figs/block-coverage.eps
	figs/block-offset.eps
	figs/block.eps
	figs/dwt.eps
	figs/frame-ordering.eps
	figs/obmc-profile.eps
	figs/sequence.eps
	figs/superblock-16pu.eps
	figs/superblock-1pu.eps
	figs/superblock-4pu.eps